Add implementation for multiply functions
This was the hardest one so far to implement, because unlike addition
and subtraction, you can't just send the numbers to the other operator
when the modifier is negative. It won't have the right answer.

This means that the division functions are going to be just as hard to
implement and that the multiplication functions could have some errors,
so I'll write some code to test the functions soon.

The division functions might not come tomorrow because it is going to
be hard to implement (what will happen when modifier is 0 (divide by 0
error), the code calling the function needs an answer). I'll probally
just commit simple bugfixes the next few days.