Remove redundancy and errors in add and subtract
The add and subtract functions for unsigned integers each had an if
statement in them that checked if they were given negative parameters.
They can only take unsigned (positive) integers. This made the if
statement useless and it is now removed.

The subtract function for unsigned integers takes a signed
modifier instead of an unsigned one. This was fixed.

The implementation for overflowDivide will come soon (by the end of the
week).