Add minor changes to non-code files
This commit didn't do much. It made the .gitignore file only ignore
files in the bin directory (where binary files will be placed). It
made LICENSE become LICENSE.txt. This is to make it more clear that it
is a text file. I also changed the README.md file to reflect this
change and fixed some spacing in there as well.