Fix errors in overflowMathTest program
When the test function in overflowMathTest was called, I forgot that it
had 6 parameters. I left out the datatype and operation paremeters.
They are added now.

Sorry for no commits in a while, I was working on other projects.
Starting to work on this again was difficult because this was not
always a library. It was just a few files in another repository I was
working on. I had to do numerous rebases and commit amending to change
all of the names and the licenses, but now it is finally done and I can
finish this project.