Fix minor errors and redundancies in test program
I had to add "#include <string>" to allow the output of std::strings
through std::cout. This was needed in the test functions to list the
variable values.

The test function declarations at the top of the file were wrong. They
didn't include the datatype or the operation variables.

Removed redundancies in the test function. The parameters did not have
to be casted to (un)signed long ints when sent to the overflowMath
functions because the test function already takes (un)signed long ints
as parameters. It was redundant and it made the code's readability
worse.

I commented out the case 2 and 3 (cases for multiply and divide) in the
test program. These functions are not implemented yet so there is no
reason for them to be in the code. Once I implement them I'll
uncomment the cases.