Fix many errors in overflowMathTest
The multiply test was not not added to the menu, this was fixed.

The menu had a redundancy in the if statement. It checked if selection
was greater than or equal to 0. It is an unsigned char, so it MUST be
greater than or equal to 0.

Multiply was not implemented in the test function. I uncommented the
lines that implemented it.

All of the test functions for the different operations were broken
because they all selected the add operation when sent to the test
function.

addTest and manualTest lacked a line at the end of the function that
tells the user to press enter to return to the menu.

There were many whitespace errors in the code. I fixed indentation and
blank lines (to separate logical parts) throughout the file.