Fix minor errors and redundancies in test program
The last commit was supposed to add the multiply function to the main
menu. It did not. This was fixed.

The "unsigned short int selection" made at the start of main() has
been changed to an unsigned char to save memory. I decided that there
most likely will not be 255+ test functions any time soon, so a char is
all we need.

There were problems with calling the overflowMath functions in the test
function. They were caused because I missed a comma when I listed the
parameters. This was fixed.

overflowMath::divide is now usable in the test function. There was no
technical reason to comment it out so one can now test it through the
manual test function.

The manual test function's menu allows the user to test multiply and
division now.

There were redundant casts in the test and manualTest functions, these
were removed.

There were missing parentheses in std::cin.ignore calls that caused
errors. The parentheses were added.