Add test function for divison
I finally added the division test function in OverflowMathTest. This
means that this library is technically done. There is still work to do,
but it is 100% usable with all of its features.

I still have to remove some redundancies in the test program. In every
function it outputs the same "Press enter" string after a test. I don't
know if I should make it a global variable or a function.

The manualTest function is extremely long (150+ lines). I will remove
redundancies and/or split it up into multiple functions.