Make main() code less clever + more user friendly
I had an if statement that took input, and subtracted it by 48 to get
the number the ASCII code represented. This was so that I could
just subtract, and save time in comparing values.

I also made the program more user friendly by waiting until the user
presses enter to end the program.