Fix extremely minor bug in main() of test program
The variable "error" wasn't declared before it was used to take input
from the user. This was fixed by just adding a single line, making this
my smallest commit.

I am sorry that so many commits have minor bugs such as these. Normally
when I write the code I'm in an environment without a compiler, so I am
unable to test the code (or even check if it compiles). I am going to
fix this by only committing after I get home to a compiler.

My next major update will most likely be to enable the library to
handle when it is sent a minNum that is larger than maxNum or vice
versa.