Add function to sort minNum and maxNum in library
I added the fixMinMax(minNum, maxNum) function to the library. This
function is called at the start of each operation function to ensure
that minNum < maxNum.

I had to choose between saving RAM memory (the size of a long int) vs
saving computation time and code complexity (adding another variable
to hold the new minNum/maxNum value). I may change this decision in the
future (OverflowMath is still in beta) if I find good reasons to do so.