Update README.md
The README.md file had lots of phrases like "this project will use" or
that the project is still in "early stages". This is no longer true so
they have been removed. I also made some general edits to the file.

The next major update will probably include a function that prevents
overflows with exponents. This can overflow easily and I am currently
trying to find the most efficient way to make this function.